Centralize equals and padBinName in abstract parent class
