+++ src/main/java/seedu/dengue/model/overview/AgeOverview.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import seedu.dengue.logic.analyst.AgeAnalyst;
-import seedu.dengue.logic.analyst.DataBin;
 import seedu.dengue.model.person.Person;
 
 /**
@@ -14,7 +13,6 @@ import seedu.dengue.model.person.Person;
  */
 public class AgeOverview extends Overview {
     private static final String AGE_TITLE = "Overview by Age";
-    private static final int MAX_AGE_LENGTH = 9; // Length of the longest bin name, "190 - 199"
 
     private AgeAnalyst analyst;
 
@@ -49,32 +47,5 @@ public class AgeOverview extends Overview {
     public void update(List<Person> personList) {
         List<Person> persons = new ArrayList<>(personList);
         this.analyst = new AgeAnalyst(persons);
-    }
-
-    private static String padBinName(String binName) {
-        int paddingNeeded = MAX_AGE_LENGTH - binName.length();
-        String whitespace = makeWhitespace(paddingNeeded / 2);
-        return whitespace + binName + whitespace;
-    }
-
-    @Override
-    public String makeBinFormat(DataBin bin) {
-        int maxSizeLen = String.valueOf(getAnalyst().getTotal()).length();
-        return String.format("%s" + GAP + "%" + maxSizeLen + "d",
-                padBinName(bin.getName()), bin.getSize());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == this) {
-            return true;
-        }
-
-        if (!(obj instanceof AgeOverview)) {
-            return false;
-        }
-
-        AgeOverview other = (AgeOverview) obj;
-        return this.analyst.equals(other.analyst);
     }
 }
