+++ src/main/java/seedu/dengue/model/overview/Overview.java
@@ -13,6 +13,7 @@ import seedu.dengue.model.person.Person;
  */
 public abstract class Overview {
     protected static final int MAX_INDEX_LEN = 4; // Length of " 28."
+    protected static final int MAX_BIN_NAME_LENGTH = 17; // Hong Leong Garden
     protected static final String GAP = "     ";
 
     /**
@@ -54,7 +55,11 @@ public abstract class Overview {
      * @param bin The input bin to format into a {@code String}.
      * @return The string representation of the bin.
      */
-    public abstract String makeBinFormat(DataBin bin);
+    public String makeBinFormat(DataBin bin) {
+        int maxSizeLen = String.valueOf(getAnalyst().getTotal()).length();
+        return String.format("%s" + GAP + "%" + maxSizeLen + "d",
+                padBinName(bin.getName()), bin.getSize());
+    }
 
     protected static String makeWhitespace(int length) {
         if (length <= 0) {
@@ -62,6 +67,12 @@ public abstract class Overview {
         }
 
         return Stream.iterate(" ", x -> x).limit(length).collect(Collectors.joining());
+    }
+
+    protected static String padBinName(String binName) {
+        int paddingNeeded = MAX_BIN_NAME_LENGTH - binName.length();
+        String whitespace = makeWhitespace(paddingNeeded / 2);
+        return whitespace + binName + whitespace;
     }
 
     private String makeBinPrefix(int x) {
@@ -81,5 +92,19 @@ public abstract class Overview {
         return Stream.iterate(1, i -> i + 1).limit(sortedBins.size())
                 .map(i -> makeBinPrefix(i) + GAP + makeBinFormat(sortedBins.get(i - 1)))
                 .collect(Collectors.joining("\n\n"));
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == this) {
+            return true;
+        }
+
+        if (!(obj instanceof Overview)) {
+            return false;
+        }
+
+        Overview other = (Overview) obj;
+        return getAnalyst().equals(other.getAnalyst());
     }
 }
