+++ src/main/java/seedu/dengue/model/overview/VariantOverview.java
@@ -3,7 +3,6 @@ package seedu.dengue.model.overview;
 import java.util.ArrayList;
 import java.util.List;
 
-import seedu.dengue.logic.analyst.DataBin;
 import seedu.dengue.logic.analyst.VariantAnalyst;
 import seedu.dengue.model.person.Person;
 
@@ -48,26 +47,5 @@ public class VariantOverview extends Overview {
     public void update(List<Person> personList) {
         List<Person> persons = new ArrayList<>(personList);
         this.analyst = new VariantAnalyst(persons);
-    }
-
-    @Override
-    public String makeBinFormat(DataBin bin) {
-        int maxSizeLen = String.valueOf(getAnalyst().getTotal()).length();
-        return String.format("%s" + GAP + "%" + maxSizeLen + "d",
-                bin.getName(), bin.getSize());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == this) {
-            return true;
-        }
-
-        if (!(obj instanceof VariantOverview)) {
-            return false;
-        }
-
-        VariantOverview other = (VariantOverview) obj;
-        return this.analyst.equals(other.analyst);
     }
 }
